refactor: Move hooks folder at src root
