+++ src/drive/web/modules/drive/Toolbar/index.jsx
@@ -6,7 +6,7 @@ import { useNavigate, useLocation, useParams } from 'react-router-dom'
 import SharingProvider, { SharedDocument } from 'cozy-sharing'
 import useBreakpoints from 'cozy-ui/transpiled/react/providers/Breakpoints'
 
-import { useCurrentFolderId } from 'drive/hooks'
+import { useCurrentFolderId } from 'hooks'
 import AddButton from 'drive/web/modules/drive/Toolbar/components/AddButton'
 import InsideRegularFolder from 'drive/web/modules/drive/Toolbar/components/InsideRegularFolder'
 import MoreMenu from 'drive/web/modules/drive/Toolbar/components/MoreMenu'
@@ -14,7 +14,7 @@ import ShareButton from 'drive/web/modules/drive/Toolbar/share/ShareButton'
 import SharedRecipients from 'drive/web/modules/drive/Toolbar/share/SharedRecipients'
 import AddMenuProvider from 'drive/web/modules/drive/AddMenu/AddMenuProvider'
 import SearchButton from 'drive/web/modules/drive/Toolbar/components/SearchButton'
-import { useDisplayedFolder } from 'drive/hooks'
+import { useDisplayedFolder } from 'hooks'
 import { useSelectionContext } from 'drive/web/modules/selection/SelectionProvider'
 import { BarRightWithProvider } from 'components/Bar'
