+++ src/drive/web/modules/drive/Toolbar/share/SharedRecipients.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { SharedRecipients } from 'cozy-sharing'
 import { useLocation, useNavigate } from 'react-router-dom'
 
-import { useDisplayedFolder } from 'drive/hooks'
+import { useDisplayedFolder } from 'hooks'
 import { getPathToShareDisplayedFolder } from 'drive/web/modules/drive/Toolbar/share/helpers'
 
 const SharedRecipientsComponent = () => {
